Clean up visuals, move location of some code
